# KAFKA-12269: Replace partition metada with auto-generated protocal