--- a/clients/src/main/java/org/apache/kafka/clients/admin/internals/MetadataOperationContext.java
+++ b/clients/src/main/java/org/apache/kafka/clients/admin/internals/MetadataOperationContext.java
@@ -17,18 +17,18 @@
 
 package org.apache.kafka.clients.admin.internals;
 
-import java.util.Collection;
-import java.util.Map;
-import java.util.Optional;
-
 import org.apache.kafka.clients.admin.AbstractOptions;
 import org.apache.kafka.common.TopicPartition;
 import org.apache.kafka.common.errors.InvalidMetadataException;
 import org.apache.kafka.common.internals.KafkaFutureImpl;
+import org.apache.kafka.common.message.MetadataResponseData;
 import org.apache.kafka.common.protocol.Errors;
 import org.apache.kafka.common.requests.MetadataResponse;
-import org.apache.kafka.common.requests.MetadataResponse.PartitionMetadata;
 import org.apache.kafka.common.requests.MetadataResponse.TopicMetadata;
+
+import java.util.Collection;
+import java.util.Map;
+import java.util.Optional;
 
 /**
  * Context class to encapsulate parameters of a call to fetch and use cluster metadata.
@@ -83,9 +83,9 @@ public final class MetadataOperationContext<T, O extends AbstractOptions<O>> {
     public static void handleMetadataErrors(MetadataResponse response) {
         for (TopicMetadata tm : response.topicMetadata()) {
             if (shouldRefreshMetadata(tm.error())) throw tm.error().exception();
-            for (PartitionMetadata pm : tm.partitionMetadata()) {
-                if (shouldRefreshMetadata(pm.error)) {
-                    throw pm.error.exception();
+            for (MetadataResponseData.MetadataResponsePartition pm : tm.partitionMetadata()) {
+                if (shouldRefreshMetadata(Errors.forCode(pm.errorCode()))) {
+                    throw Errors.forCode(pm.errorCode()).exception();
                 }
             }
         }