… clients/src/test/java/org/apache/kafka/clients/MetadataCacheTest.java
… clients/src/test/java/org/apache/kafka/clients/MetadataTest.java
… clients/src/test/java/org/apache/kafka/clients/admin/KafkaAdminClientTest.java
… clients/src/test/java/org/apache/kafka/clients/consumer/internals/ConsumerCoordinatorTest.java
… clients/src/test/java/org/apache/kafka/clients/consumer/internals/ConsumerMetadataTest.java
… clients/src/test/java/org/apache/kafka/clients/consumer/internals/FetcherTest.java
… clients/src/test/java/org/apache/kafka/common/requests/RequestResponseTest.java
… clients/src/test/java/org/apache/kafka/common/requests/RequestTestUtils.java
… core/src/test/scala/unit/kafka/admin/AddPartitionsTest.scala
… core/src/test/scala/unit/kafka/server/MetadataRequestTest.scala
… core/src/test/scala/unit/kafka/server/MetadataRequestWithForwardingTest.scala
… jmh-benchmarks/src/main/java/org/apache/kafka/jmh/metadata/MetadataResponseBenchmark.java