*More detailed description of your change*
MetadataResponse.PartitionMetadata is almost the same as MetadataResponsePartition so we can replace it, below is the difference that we need to notice:
1. `MetadataResponse.PartitionMetadata` contains a `TopicPartition` but `MetadataResponsePartition` only contains partitionIndex, so we need to change some method to add topic if necessary.
2. leaderId and leaderEpoch in `MetadataResponse.PartitionMetadata` are `Optional`, we use `NO_LEADER_ID` and `NO_PARTITION_LEADER_EPOCH` to replace null `Optional`.

Following work:
The `MetadataResponse.TopicMetadata` can also be replaced in a next pr, this pr only replace `MetadataResponse.PartitionMetadata` to make pr simple for review.

*Summary of testing strategy (including rationale)*
QA

### Committer Checklist (excluded from commit message)
- [ ] Verify design and implementation 
- [ ] Verify test coverage and CI build status
- [ ] Verify documentation (including upgrade notes)