@ijuma 2021-02-03
Thanks for the PR. Let's run the relevant jmh benchmark to ensure there is no perf regression.